Update command processor to parse arguments and add dummy commands
I'm working on adding dynamically updatable deadline and lecture records
for Tri 2. I decided to start with the front-end.

This commit implements the following things:

* [comproc] Argument parsing
* [comproc] Automated help pages (for admins)
* [comproc] Dummy commands for manipulating courses, deadlines, lectures
* [data] JSON data files replacing remaining CSV data files
* [data] Removal of passed Tri 1 courses and lectures
* [courses] Started work on new module to replace `class-times.js`

This is the first commit of the sprint I'm going to try do over the two
week break.

After this is done and sorted with, this project will be containerised
and have CI added.